Update sentence_translations.json
There is a major case of diglossia among people that know Hindi, which is why I have incorporated multiple words of the same meaning at various junctures to keep any potential listener in the know. So, if you run these translations through a translator, you might see the same word appearing twice in a sentence for this very reason.