CLT Hebrew subtitles update
Updated Hebrew subtitles for the CTL lecture.  Since I don't know how to view the lecture using this subtitles version, I don't know how well it is showing.  This is important, since there are some major issues when editing mixed RTL (Hebrew) and LTL (English, numbers) text.

There is one issue that needs discussion - male and female language.  I believe this is a general issue for other languages like Arabic.  The problem is that in most cases the auto translation of "you" is to the male variant.  In Hebrew, "you" can be single male/female (I will call it "youM"/"youF") and plural male/female ("youPM"/"youPF").

One way to fix that is to translate "you can ..." to "youF/M canF/M ...", but this is too cumbersome.  Another option, which I used in this translation is to change to plural male - "youPM canPM ...".  This is usually acceptable also for women.

A third option is to remove the direct reference to the viewers/students. E.g "you can do something" becomes "something can be done".  This completely solves the mail/female issue, but I don't know if this is an acceptable approach, because of the removal of the reference to the viewers.

Your thoughts about this issue can help.