Update sentence_translations.json
Following the initial translation by translating  any "they" which stands for a single person into the male form. Sadly, German language has no gender-neutral counterpart.
Maybe this should be changed into all-female just for fun. I don't know.

Not sure about the correct translation of the word "reprint" in Wigners's story. Is it supposed to be a reprint of one of the statistician's own works, or some standard text book (or sth else entirely?)

One remark why a different interface might be helpful after all (but not using the GitHub interface for editing normally, so I'm not sure...): If I edit on the web interface, it doesn't store intermediate editing steps, right? Because this here took me several hours, and usually I just cannot do that in one run, however commits of partial work are also really unsatisfactory.