CLT Hebrew subtitles update#2
Enhancing the previous enhancement to the CTL Hebrew subtitles, after viewing the video with the updated subtitles.  It would really help if there will be a way to view the updated subtitles with the video before committing the PR.

Also, I noticed in one case that the English text used for the Hebrew translation was not the same as the English subtitles for the video.  If it is possible to translate the text only after the English text is final, it would help for both the auto translation and the updates to the translation.