Update sentence_translations.json
In my opinion a lot of Indians (including me) aren't that comfortable with reading Hindi captions in Devanagari (script of Hindi) because even when we type something in Hind we type the Hindi words in Roman script, and the only time we ever actually write Hindi is in school, which is also only till tenth grade for most students as many students don't take Hindi as one of their classes in  eleventh grade. So my suggestion is that you add a Hindi text written in Roman script (like I did here) to appeal to a larger audience or create a separate section like "Hinglish" for this purpose.