Add support for RTL languages
RTL languages include: Hebrew, Arabic, Persian and Urdu (let me know if
I missed any)

The text area for right to left languages needs special styling
(`direction: rtl`), otherwise editing sentences in these languages is a
rather miserable experience (especially when mixing numbers / English
phrases which are written left to write).

Also the "legacy translation" div needs to be right to left as well. And
its short label has to be translated to the respective language, since
its concatenated on the right, and it looks awkward when it's in
English.

I've relied on Google Translate / copliot to translate the "legacy
translation" label to Arabic, Persian and Urdu, so I'm not actually sure
they're accurate or make any sense, feel free to offer alternative
phrasing.